requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0